FIX : 스크린들 디렉터리로 리팩토링
