+++ front/src/screens/Auths/SplashScreen.tsx
