+++ front/src/screens/Storybooks/ImageEditorScreen.tsx
